+++ state/src/views/state/s2s/common/bridge-skeleton.vue
